+++ pkg/provider/provider.go
@@ -194,6 +194,9 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 		NewNotificationPolicyResource,
 		NewNotebookResource,
 		NewAppOAuthResource,
+		NewUserMCPToolMetadataResource,
+		NewUserMCPPromptMetadataResource,
+		NewUserMCPResourceMetadataResource,
 	}
 }